+++ anki_lexique_flashcard_generator/Easy_Run_s1_s7.py
@@ -9,7 +9,7 @@
 from anki_lexique_flashcard_generator.s1_Filter_Lexique import clean_lexique as s1_clean_lexique
 from anki_lexique_flashcard_generator.s2_Mux_Lexique import CHUNK_SIZE
 from anki_lexique_flashcard_generator.s3_Export_Anki_Format import export_anki_format_csvs as s3_export_anki_format_csvs
-from anki_lexique_flashcard_generator.s4_Enforce_Num_Anki_Cards import enforce_chunk_size_anki_cards as s4_enforce_chunk_size_anki_cards
+from anki_lexique_flashcard_generator.s4_Enforce_Num_CSV_Rows import enforce_chunk_size_anki_cards as s4_enforce_chunk_size_anki_cards
 from anki_lexique_flashcard_generator.s5_Generate_Anki_Package import generate_anki_packages as s5_generate_anki_packages
 from anki_lexique_flashcard_generator.s6_Import_Packages_Into_Anki import import_anki_packages as s6_import_anki_packages
 from anki_lexique_flashcard_generator.s7_Organize_Deck import organize_deck as s7_organize_deck
